requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.2.1
django-celery-beat==1.4.0
django-timezone-field==3.0
djangorestframework==3.9.4
kombu==4.5.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
